refactor: keep install command under lint cap
